requirements.txt
langchain
black
python-dotenv
openai
tweepy
flask
typing
google-search-results